requirements.txt
asttokens==2.2.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
executing==1.2.0
Flask==2.2.2
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
idna==3.4
importlib-metadata==5.1.0
ipdb==0.13.9
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
numpy==1.23.5
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.33
protobuf==4.21.11
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
rsa==4.9
six==1.16.0
stack-data==0.6.2
toml==0.10.2
traitlets==5.6.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.11.0
